Diagramm values as floats

+++ Arduino_Pedelec_Controller/Display/Components.h
@@ -40,7 +40,7 @@ typedef struct _View {
     int8_t active_components_ids[MAX_COMP_ACTIVE];
     String diagram_string;
     ValueId diagram_val;
-    uint8_t diagram_precision;
+    int16_t diagram_precision;
     float_t diagram_min;
     float_t diagram_max;
 } View;
